refactor(dashboard): split dashboard grid dll
