+++ redux/action/fisrtLoad.action.js
@@ -1,0 +1,13 @@
+import {SET_FIRST_LOAD_FALSE,SET_FIRST_LOAD_TRUE} from '../constan/index'
+
+export const setFisrtLoadFalse = () => {
+  return {
+    type: SET_FIRST_LOAD_FALSE
+  }
+}
+
+export const setFisrtLoadTrue = () => {
+  return {
+    type: SET_FIRST_LOAD_TRUE
+  }
+}
