+++ redux/constan/index.js
@@ -7,3 +7,6 @@ export const LOGIN_LOADING = 'LOGIN_LOADING';
 export const LOGIN_SUCCESS = 'LOGIN_SUCCESS';
 export const LOGIN_FAILED = 'LOGIN_FAILED';
 export const LOGIN_DEFAULT = 'LOGIN_DEFAULT';
+
+export const SET_FIRST_LOAD_FALSE = 'SET_FIRST_LOAD_FALSE'
+export const SET_FIRST_LOAD_TRUE = 'SET_FIRST_LOAD_TRUE'
