+++ redux/reducer/index.js
@@ -1,8 +1,10 @@
 import { combineReducers } from "redux"
 import UserReducer from "./user/userReducer"
 import LoginReducer from "./user/loginReducer"
+import FirstLoad from '../reducer/firstLoad.reducer'
 
 export default combineReducers({
     UserReducer,
-    LoginReducer
+    LoginReducer,
+    FirstLoad,
 })
